Allow comparison of spans only differing in cv-ness
Note: The C++ standard does not allow to compare containers that only differ in cv-ness.

This also holds for M-GSL. However M-GSL does allow to compare string_span with cstring_span.

The comparation of string_span in GSL Lite (as it is now) is the same as that of span.
To continue allowing to compare string_span with cstring_span in GSL Lite, it is now possible to compare spans differing in cv-ness.

Perhaps a configuration flag that allows to switch between this behavior and the strict behavior is in order.