`gsl_noexcept` no longer depends on contract violation settings
Having the definition of `gsl_noexcept` vary depending on a
configuration macro can lead to type and ABI incompatibilities because
`noexcept` is part of the type system since C++17.

Following the Lakos rule, functions with a narrow contract should not be
marked `noexcept`, not even conditionally.
(cf. https://quuxplusone.github.io/blog/2018/04/25/the-lakos-rule/)

An exception is made for functions which only call move constructors or
move assignment operators; these are made unconditionally `noexcept`.
The reasoning is that a throwing move constructor is unsound anyway.