Fix #201: `ExpectsAudit()` should have `gsl_` prefix (#206)
* Rename `ExpectsAudit()` to `gsl_ExpectsAudit()`

Reason: it's an extension and not a feature specified in the Core
Guidelines. See #201.

* Restructure contract configuration macros

We now have the following independent groups:

- contract checking control:
  - `gsl_CONFIG_CONTRACT_CHECKING_AUDIT`
  - `gsl_CONFIG_CONTRACT_CHECKING_ON` (default)
  - `gsl_CONFIG_CONTRACT_CHECKING_OFF`
- contract checking control for individual contract types:
  - `gsl_CONFIG_CONTRACT_CHECKING_EXPECTS_OFF`
  - `gsl_CONFIG_CONTRACT_CHECKING_ENSURES_OFF`
- contract violation control:
  - `gsl_CONFIG_CONTRACT_VIOLATION_THROWS`
  - `gsl_CONFIG_CONTRACT_VIOLATION_TERMINATES` (default)
  - `gsl_CONFIG_CONTRACT_VIOLATION_CALLS_HANDLER`
- how unenforced contracts are handled:
  - `gsl_CONFIG_UNENFORCED_CONTRACTS_ASSUME`
  - `gsl_CONFIG_UNENFORCED_CONTRACTS_ELIDE` (default)

* `gsl_noexcept` no longer depends on contract violation settings

Having the definition of `gsl_noexcept` vary depending on a
configuration macro can lead to type and ABI incompatibilities because
`noexcept` is part of the type system since C++17.

Following the Lakos rule, functions with a narrow contract should not be
marked `noexcept`, not even conditionally.
(cf. https://quuxplusone.github.io/blog/2018/04/25/the-lakos-rule/)

An exception is made for functions which only call move constructors or
move assignment operators; these are made unconditionally `noexcept`.
The reasoning is that a throwing move constructor would be unsound
anyway.