Add missing checks in `not_null<>` move and copy constructors
We had non-null checks in converting constructors but not in plain
move/copy constructors. We want to be consistent and have checks
everywhere.

Also, pre- and postconditions are now expressed with `Expects()` and
`Ensures()` directly in the functions that impose them rather than
being delegated to `checked_ptr()`. This is preferable because of the
self-documenting nature of contract checks, and because it is closer to
the Contracts draft language feature which makes contract checks part
of the function signature.

(cherry picked from commit 39849638e4287790170af97e3176fbde0fbb9e2b)