Unify and document SFINAE conventions; add missing checks in not_null<> move and copy constructors (#208)
* Unify and document our SFINAE conventions

(cherry picked from commit c32e88cff58764b9f970489f03f525dd844e7875)

* Add missing checks in `not_null<>` move and copy constructors

We had non-null checks in converting constructors but not in plain
move/copy constructors. We want to be consistent and have checks
everywhere.

Also, pre- and postconditions are now expressed with `Expects()` and
`Ensures()` directly in the functions that impose them rather than
being delegated to `checked_ptr()`. This is preferable because of the
self-documenting nature of contract checks, and because it is closer to
the Contracts draft language feature which makes contract checks part
of the function signature.

(cherry picked from commit 39849638e4287790170af97e3176fbde0fbb9e2b)

* Test propagation failure of moved-from `not_null<>`

* Add type trait tests for `not_null<>`; fix `is_copy_constructible<>`

We don't have SFINAE in copy constructors and copy assignment operators,
so we use a base class with explicit specializations to work around.

* Fix SFINAE in conditionally explicit `not_null<>` conversion constructor

* Add additional test targets with alternative configurations

* Rename `gsl_CONFIG_NOT_NULL_TRANSPARENT_GET` to `*_TRANSPARENT_NOT_NULL`

* Add Xcode 10.3 and 11.2 to Travis configuration

* Try to refine Apple Clang version detection

* Work around Apple Clang errors in source, not in .travis.yml

* VS 2013 supports deleted functions

* Add explicit workaround for `unique_ptr<>` on VC++ 2013

...to work around the buggy traits implementation.

* Disable some trait tests unless `gsl_CONFIG_NOT_NULL_EXPLICIT_CTOR=1`

* Make `not_null_data<>` a member, not a base class

Also, checks of the class invariant now use `Ensures()` rather than
`Expects()`, and fix some bugs in the assignment operator checks.