Use `gsl_Assert()` for invariant checks; fix `weak_ptr<>` conversion (#311)
The `not_null<>` conversion operators previously made the unfounded and unnecessary assumption that the type being converted to is also pointer-like. This was a problem when converting `not_null<shared_ptr<T>>` to `weak_ptr<T>`, which is not pointer-like despite the name.

Also, the use of `gsl_Ensures()` in the member functions of `not_null<>` was misguided; these checks express class invariants, not postconditions, and should therefore be written using `gsl_Assert()`.