Fix `not_null_ic<>` copy construction (#326, thanks to @runer112)
The conversion constructor should accept its argument by value to avoid dreaded interference with the copy and move constructors. With a forwarding reference in the conversion constructor, the copy and move constructors are overridden in some scenarios, ending up with undefined behavior (by evaluating type traits for an incomplete type) and, in this case, causing infinite recursion at runtime.