Add `span<>` constructor overloads for iterator arguments
This improves backward compatibility, since our iterator used to be a pointer
in v0.*, which made `span{ otherSpan.begin(), 0 }` work because we have a
pointer overload.

We don't add a constrained overload for contiguous overloads because that
would require not only C++20 but also the `<iterator>` header for the
`contiguous_iterator<>` concept, which I'd prefer not to pull in. Providing
overloads for our own iterator seems like a good compromise here.